Numeric.Natural

Description

Natural numbers.

The Numeric.Natural module has become part of base starting with `base-4.8.0.0`.

Documentation

data Natural

Type representing arbitrary-precision non-negative integers.

Operations whose result would be negative throw (Underflow :: ArithException).

The Natural type has become part of base starting with `base-4.8.0.0`.