Numeric.Natural.Internal

Description

This module exposes the potentially unsafe operations that are sometimes needed for efficiency: The Natural data constructor and unsafePred.

Documentation

newtype Natural

Constructors

Natural

Fields

runNatural :: Integer

class Integral n => Whole n where

A refinement of Integral to represent types that do not contain negative numbers.

Methods

toNatural :: n -> Natural

unsafePred :: n -> n

natural :: a -> (a -> a) -> Natural -> a

Church decoding